fix: move strategy selection into strategies screen
